// src/shared/models/common/enums.ts
export enum QuestionTypeEnum {
  MULTIPLE_CHOICE = 'MULTIPLE_CHOICE',
  TEXT = 'TEXT',
  CODE = 'CODE',
  BOOLEAN = 'BOOLEAN',
}
// ...
export enum JobAiAssessmentResultEnum {
  NOT_AVAILABLE = 'NOT_AVAILABLE',
  PASSED = 'PASSED',
  AI_REVIEW_FAILED = 'AI_REVIEW_FAILED',
  MANUAL_REVIEW_FAILED = 'MANUAL_REVIEW_FAILED',
}
// ...
export enum JobAiAssessmentRecommendationEnum {
  HIGHLY_RECOMMENDED = 'HIGHLY_RECOMMENDED',
  RECOMMENDED = 'RECOMMENDED',
  NOT_RECOMMENDED = 'NOT_RECOMMENDED',
}
// ...
export enum JobAiAssessmentStatusEnum {
  NOT_STARTED = 'NOT_STARTED',
  AI_INITIALIZATION_IN_PROGRESS = 'AI_INITIALIZATION_IN_PROGRESS',
  AI_INITIALIZATION_COMPLETED = 'AI_INITIALIZATION_COMPLETED',
  CANDIDATE_ASSESSMENT_IN_PROGRESS = 'CANDIDATE_ASSESSMENT_IN_PROGRESS',
  CANDIDATE_ASSESSMENT_COMPLETED = 'CANDIDATE_ASSESSMENT_COMPLETED',
  AI_REVIEW_IN_PROGRESS = 'AI_REVIEW_IN_PROGRESS',
  AI_REVIEW_COMPLETED = 'AI_REVIEW_COMPLETED',
  MANUAL_REVIEW_IN_PROGRESS = 'MANUAL_REVIEW_IN_PROGRESS',
  MANUAL_REVIEW_COMPLETED = 'MANUAL_REVIEW_COMPLETED',
  ASSESSMENT_COMPLETED = 'ASSESSMENT_COMPLETED',
  ASSESSMENT_FAILED = 'ASSESSMENT_FAILED',
}

/**
 * @openapi
 * components:
 *   schemas:
 *     JobAiAssessmentSectionStatusEnum:
 *       type: string
 *       description: Status of an JobAi assessment section
 *       enum:
 *         - NOT_STARTED
 *         - IN_PROGRESS
 *         - COMPLETED
 *         - FAILED
 */
export enum JobAiAssessmentSectionStatusEnum {
  NOT_STARTED = 'NOT_STARTED',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
}

/**
 * @openapi
 * components:
 *   schemas:
 *     JobAiAssessmentSectionResultEnum:
 *       type: string
 *       description: Result of an JobAi assessment section
 *       enum:
 *         - NOT_AVAILABLE
 *         - PASSED
 *         - FAILED_AI_REVIEW
 *         - FAILED_MANUAL_REVIEW
 */
export enum JobAiAssessmentSectionResultEnum {
  NOT_AVAILABLE = 'NOT_AVAILABLE',
  PASSED = 'PASSED',
  FAILED_AI_REVIEW = 'FAILED_AI_REVIEW',
  FAILED_MANUAL_REVIEW = 'FAILED_MANUAL_REVIEW',
}

/**
 * @openapi
 * components:
 *   schemas:
 *     JobAiAssessmentTaskStatusEnum:
 *       type: string
 *       description: Status of an JobAi assessment initialize task
 *       enum:
 *         - PENDING
 *         - INITIALIZE_STARTED
 *         - INITIALIZE_COMPLETED
 *         - ASSESSMENT_STARTED
 *         - ASSESSMENT_COMPLETED
 *         - COMPLETED
 *         - FAILED
 */
export enum JobAiAssessmentTaskStatusEnum {
  PENDING = 'PENDING',
  INITIALIZE_STARTED = 'INITIALIZE_STARTED',
  INITIALIZE_COMPLETED = 'INITIALIZE_COMPLETED',
  ASSESSMENT_STARTED = 'ASSESSMENT_STARTED',
  ASSESSMENT_COMPLETED = 'ASSESSMENT_COMPLETED',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
}